# altpro-api/src/main/java/com/taurasg/altpro/api/model/Organization.java
package com.taurasg.altpro.api.model;

import jakarta.validation.constraints.NotBlank;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;
import org.springframework.data.annotation.Id;
import org.springframework.data.mongodb.core.mapping.Document;

import java.time.Instant;
import java.util.List;

@Document(collection = "organizations")
@Data
@NoArgsConstructor
@AllArgsConstructor
public class Organization {
    @Id
    private String id;

    @NotBlank(message = "name is required")
    private String name;

    private String description;
    private Instant createdAt;

    private List<OrgMember> members;
}